X axis ticks only painted for number of selected x values
Until now all ticks have been painted which does not makes sense if you have thousands of values and only want to show 10. the svg path for the tick mark is now only painted if the corresponding value is also painted. thanks for your fork!

---
This commit is on behalf of @gerbsen <https://github.com/gerbsen> who had submitted this patch for an older branch (which I had messed up). Thanks gerbsen.